GaFi_Prediction/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pandas
numpy
scikit-learn
prophet
joblib
pydantic
httpx
matplotlib
seaborn
jupyter
ipykernel
